+++ test/types/helpers.test-d.ts
@@ -36,6 +36,8 @@ import {
   ScrollSearchResponse,
   OnDropDocument,
   MsearchHelper,
+  UpdateActionDocOperation,
+  UpdateActionScriptOperation,
 } from '../../lib/Helpers';
 import { ApiResponse, ApiError, Context } from '../../lib/Transport';
 
@@ -107,7 +109,20 @@ expectError(
   const options: BulkHelperOptions<Record<string, any>> = {
     datasource: [],
     onDocument(doc: Record<string, any>) {
-      return [{ update: { _index: 'test' } }, doc];
+      return [{ update: { _index: 'test' } }, doc as UpdateActionDocOperation];
+    },
+  };
+  expectAssignable<BulkHelperOptions<Record<string, any>>>(options);
+}
+// update
+{
+  // without `:BulkHelperOptions` this test cannot pass
+  // but if we write these options inline inside
+  // a `.helper.bulk`, it works as expected
+  const options: BulkHelperOptions<Record<string, any>> = {
+    datasource: [],
+    onDocument(doc: Record<string, any>) {
+      return [{ update: { _index: 'test' } }, doc as UpdateActionScriptOperation];
     },
   };
   expectAssignable<BulkHelperOptions<Record<string, any>>>(options);